requirements.txt
uwsgi==2.0.17
Django==3.2.5
pytest-django==4.4.0
djangorestframework==3.12.4
django-log-request-id==1.3.2
drf-spectacular==0.17.3
requests==2.26.0